# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FB Stop Bot API",
    "description": "Минимальный набор endpoints после миграции.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/dashboard/ads": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "List Ad Snapshots",
        "description": "Список ad'ов с alert_state, последней метрикой и offer.\n\nПартиционный фильтр по ad_metrics.cycle_ts применяется внутри LATERAL'а\n(последние 7 дней). Если за окно нет метрик — metrics=None.\n\nX-Total-Count в headers — реальный COUNT с теми же фильтрами без LIMIT,\nдля пагинации фронта.",
        "operationId": "list_ad_snapshots_api_dashboard_ads_get",
        "parameters": [
          {
            "name": "alert_state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "CSV alert_state'ов (warning_sent,stop_sent,...)",
              "title": "Alert State"
            },
            "description": "CSV alert_state'ов (warning_sent,stop_sent,...)"
          },
          {
            "name": "fb_ad_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "CSV Meta-ID объявлений",
              "title": "Fb Ad Ids"
            },
            "description": "CSV Meta-ID объявлений"
          },
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inactive"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdSnapshotOut"
                  },
                  "title": "Response List Ad Snapshots Api Dashboard Ads Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/alerts": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "List Alert Events",
        "description": "Append-only лог FSM-событий с JOIN'ами по fb_ads/fb_campaigns/offers.\n\nПартиционный фильтр по alert_events.created_at — обязательный.\nЕсли from_iso/to_iso не переданы — дефолт last 24h.\nЕсли from_iso > to_iso → 422.\n\nCRITICAL: имена полей AlertEvent — `stage` / `matched_rule_codes`\n(не `event_type` / `rule_codes`). triggered_by_rule_codes не существует\nв ORM — возвращаем None в ответе для совместимости с frontend shape.",
        "operationId": "list_alert_events_api_dashboard_alerts_get",
        "parameters": [
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "warning | stop",
              "title": "Stage"
            },
            "description": "warning | stop"
          },
          {
            "name": "fb_ad_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Meta-ID объявления",
              "title": "Fb Ad Id"
            },
            "description": "Meta-ID объявления"
          },
          {
            "name": "from_iso",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 начало окна",
              "title": "From Iso"
            },
            "description": "ISO-8601 начало окна"
          },
          {
            "name": "to_iso",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 конец окна",
              "title": "To Iso"
            },
            "description": "ISO-8601 конец окна"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlertEventOut"
                  },
                  "title": "Response List Alert Events Api Dashboard Alerts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/incidents": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "List Incidents",
        "description": "Активные инциденты: ad_alert_state IN ('warning_sent','stop_sent')\nAND (snoozed_until IS NULL OR snoozed_until < NOW()).\n\nДополнительно к /dashboard/ads:\n- incident_open_since: last_transition_at (момент входа в текущее состояние).\n- incident_duration_seconds: NOW - incident_open_since.\n- transitions_count: COUNT(alert_events) от incident_open_since до NOW().\n\ntransitions_count считается batch'ем за один запрос — никаких N+1.",
        "operationId": "list_incidents_api_dashboard_incidents_get",
        "parameters": [
          {
            "name": "stage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "warning | stop | all",
              "default": "all",
              "title": "Stage"
            },
            "description": "warning | stop | all"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncidentOut"
                  },
                  "title": "Response List Incidents Api Dashboard Incidents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdSnapshotOut": {
        "properties": {
          "fb_ad_id": {
            "type": "string",
            "title": "Fb Ad Id"
          },
          "internal_id": {
            "type": "string",
            "title": "Internal Id"
          },
          "ad_name": {
            "type": "string",
            "title": "Ad Name"
          },
          "campaign_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Name"
          },
          "adset_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adset Name"
          },
          "offer_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offer Code"
          },
          "offer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offer Id"
          },
          "alert_state": {
            "type": "string",
            "title": "Alert State",
            "default": "normal"
          },
          "snoozed_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snoozed Until"
          },
          "open_state_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open State Token"
          },
          "last_warning_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Warning At"
          },
          "last_stop_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Stop At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "delivery_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Status"
          },
          "meta_ad_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Ad Status"
          },
          "stop_rule_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Stop Rule Codes"
          },
          "warning_rule_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warning Rule Codes"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MetricsBlock"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "fb_ad_id",
          "internal_id",
          "ad_name",
          "is_active"
        ],
        "title": "AdSnapshotOut",
        "description": "Композитный snapshot одного ad'а для /dashboard/ads и /dashboard/incidents."
      },
...
      "AlertEventOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "fb_ad_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fb Ad Id"
          },
          "ad_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ad Name"
          },
          "campaign_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Name"
          },
          "offer_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offer Code"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "matched_rule_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Rule Codes"
          },
          "triggered_by_rule_codes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Triggered By Rule Codes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "alert_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Payload"
          }
        },
        "type": "object",
        "required": [
          "id",
          "stage",
          "created_at"
        ],
        "title": "AlertEventOut",
        "description": "Одна запись alert_events с JOIN'ом по fb_ads/offers."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IncidentOut": {
        "properties": {
          "fb_ad_id": {
            "type": "string",
            "title": "Fb Ad Id"
          },
          "internal_id": {
            "type": "string",
            "title": "Internal Id"
          },
          "ad_name": {
            "type": "string",
            "title": "Ad Name"
          },
          "campaign_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Name"
          },
          "adset_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adset Name"
          },
          "offer_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offer Code"
          },
          "offer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offer Id"
          },
          "alert_state": {
            "type": "string",
            "title": "Alert State",
            "default": "normal"
          },
          "snoozed_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snoozed Until"
          },
          "open_state_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open State Token"
          },
          "last_warning_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Warning At"
          },
          "last_stop_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Stop At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "delivery_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Status"
          },
          "meta_ad_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Ad Status"
          },
          "stop_rule_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Stop Rule Codes"
          },
          "warning_rule_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warning Rule Codes"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MetricsBlock"
              },
              {
                "type": "null"
              }
            ]
          },
          "incident_open_since": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Open Since"
          },
          "incident_duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Duration Seconds"
          },
          "transitions_count": {
            "type": "integer",
            "title": "Transitions Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "fb_ad_id",
          "internal_id",
          "ad_name",
          "is_active"
        ],
        "title": "IncidentOut",
        "description": "AdSnapshotOut + incident-специфичные поля для /dashboard/incidents."
      },
...
      "MetricsBlock": {
        "properties": {
          "cycle_ts": {
            "type": "string",
            "title": "Cycle Ts"
          },
          "spend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spend"
          },
          "impressions": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impressions"
          },
          "clicks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clicks"
          },
          "ctr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ctr"
          },
          "cpc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cpc"
          },
          "cpm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cpm"
          },
          "reach": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reach"
          },
          "frequency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frequency"
          },
          "leads": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leads"
          },
          "cost_per_lead": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Per Lead"
          },
          "registrations": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registrations"
          },
          "cost_per_registration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Per Registration"
          },
          "deposits": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deposits"
          }
        },
        "type": "object",
        "required": [
          "cycle_ts"
        ],
        "title": "MetricsBlock",
        "description": "Блок метрик (последняя строка ad_metrics за окно 7 дней).\n\nDecimal'ы сериализуются как str — стабильный путь без потери точности."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}